+++ sqllin-dsl/src/commonMain/kotlin/com/ctrip/sqllin/dsl/sql/clause/Function.kt
@@ -16,6 +16,7 @@
 
 package com.ctrip.sqllin.dsl.sql.clause
 
+import com.ctrip.sqllin.dsl.annotation.FunctionDslMaker
 import com.ctrip.sqllin.dsl.sql.Table
 import com.ctrip.sqllin.dsl.sql.X
 
@@ -24,32 +25,42 @@ import com.ctrip.sqllin.dsl.sql.X
  * @author yaqiao
  */
 
+@FunctionDslMaker
 public fun <T> Table<T>.count(element: ClauseElement): ClauseNumber =
     ClauseNumber("count(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.count(x: X): ClauseNumber =
     ClauseNumber("count(*)", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.max(element: ClauseElement): ClauseNumber =
     ClauseNumber("max(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.min(element: ClauseElement): ClauseNumber =
     ClauseNumber("min(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.avg(element: ClauseElement): ClauseNumber =
     ClauseNumber("avg(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.sum(element: ClauseElement): ClauseNumber =
     ClauseNumber("sum(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.abs(number: ClauseElement): ClauseNumber =
     ClauseNumber("abs(${number.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.upper(element: ClauseElement): ClauseString =
     ClauseString("upper(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.lower(element: ClauseElement): ClauseString =
     ClauseString("lower(${element.valueName})", this, true)
 
+@FunctionDslMaker
 public fun <T> Table<T>.length(element: ClauseElement): ClauseNumber =
     ClauseNumber("length(${element.valueName})", this, true)