+++ sqllin-dsl/src/commonMain/kotlin/com/ctrip/sqllin/dsl/sql/clause/LeftOuterJoinClause.kt
@@ -16,6 +16,7 @@
 
 package com.ctrip.sqllin.dsl.sql.clause
 
+import com.ctrip.sqllin.dsl.annotation.StatementDslMaker
 import com.ctrip.sqllin.dsl.sql.Table
 
 /**
@@ -30,6 +31,7 @@ internal class LeftOuterJoinClause<R>(
     override val clauseName: String = " LEFT OUTER JOIN "
 }
 
+@StatementDslMaker
 public fun <R> LEFT_OUTER_JOIN(vararg tables: Table<*>): JoinClause<R> = LeftOuterJoinClause(*tables)
 
 internal class NaturalLeftOuterJoinClause<R>(
@@ -39,4 +41,5 @@ internal class NaturalLeftOuterJoinClause<R>(
     override val clauseName: String = " NATURAL LEFT OUTER JOIN "
 }
 
+@StatementDslMaker
 public fun <R> NATURAL_LEFT_OUTER_JOIN(vararg tables: Table<*>): NaturalJoinClause<R> = NaturalLeftOuterJoinClause(*tables)